[Feat] 책 상태 선택기능 구현
