+++ src/components/atoms/SaveButton/SaveButton.jsx
@@ -1,5 +1,13 @@
 function SaveButton() {
-  return <button className="contents-md-bold text-primary-500">저장</button>;
+  return (
+    <button
+      type="submit"
+      form="bookInfo"
+      className="contents-md-bold text-primary-500"
+    >
+      저장
+    </button>
+  );
 }
 
 export default SaveButton;
