+++ src/components/molecules/BookInfoState/BookInfoState.jsx
@@ -1,43 +1,29 @@
 import PropTypes from 'prop-types';
 import BookInfoStateButton from '../../atoms/BookInfoStateButton/BookInfoStateButton';
-import { useState } from 'react';
-import BookInfoDetail from '../BookInfoDetail/BookInfoDetail';
 
-function BookInfoState() {
-  const [status, setStatus] = useState('완독');
-
-  const handleClick = (e) => {
-    e.preventDefault();
-    const button = e.target.closest('button');
-    if (!button) return;
-
-    setStatus(button.innerText);
-  };
-
+function BookInfoState({ status, onClick }) {
   return (
     <>
       <div className="flex justify-center gap-2 items-center">
-        <BookInfoStateButton
-          title={'완독'}
-          onClick={handleClick}
-          status={status}
-        />
+        <BookInfoStateButton title={'완독'} onClick={onClick} status={status} />
         <BookInfoStateButton
           title={'독서중'}
-          onClick={handleClick}
+          onClick={onClick}
           status={status}
         />
         <BookInfoStateButton
           title={'희망 도서'}
-          onClick={handleClick}
+          onClick={onClick}
           status={status}
         />
       </div>
-      <BookInfoDetail status={status} />
     </>
   );
 }
 
-BookInfoState.propTypes = {};
+BookInfoState.propTypes = {
+  status: PropTypes.string.isRequired,
+  onClick: PropTypes.func.isRequired,
+};
 
 export default BookInfoState;
